--- src/ability/mod.rs
@@ -1,7 +1,0 @@
-use bevy::prelude::*;
-
-pub struct AbilityPlugin;
-
-impl Plugin for AbilityPlugin {
-    fn build(&self, app: &mut App) {}
-}
