+++ src/ai/simple_ai.rs
@@ -4,15 +4,11 @@ use iyes_loopless::prelude::*;
 use pino_utils::{ok_or_continue, ok_or_return};
 
 use crate::{
-    attack::AttackEvent,
     enemy::pathfinding::a_star,
     game::GameState,
-    grid::{CellType, Grid, GridEntity},
-    movement::Movement,
     player::Player,
-    spritesheet::SpriteFrames,
+    prelude::*,
     ui::{attack_animation::SpawnAttackAnimEvent, attack_indicator::AttackIndicator},
-    utils::Dir,
 };
 
 /// Track distance to the player
