+++ src/enviro/arrow_trap.rs
@@ -3,16 +3,8 @@ use bevy_ecs_ldtk::{prelude::FieldValue, EntityInstance};
 use iyes_loopless::prelude::*;
 
 use crate::{
-    attack::AttackEvent,
-    constants::BEING_LAYER,
-    game::GameState,
-    grid::{to_world_coords, CellType, GridEntity},
-    map::ldtk_to_bevy,
-    screens::state::ScreenState,
-    spritesheet::SpriteSheet,
+    prelude::*,
     ui::{attack_indicator::AttackIndicator, projectile::SpawnProjectileEvent},
-    utils::{cleanup, Dir},
-    weapon::CurrentWeapon,
 };
 
 // how many turns between each attack
