+++ src/enviro/collapsable_floor.rs
@@ -3,15 +3,7 @@ use bevy_bobs::component::health::Health;
 use bevy_ecs_ldtk::EntityInstance;
 use iyes_loopless::prelude::*;
 
-use crate::{
-    constants::BEING_LAYER,
-    game::GameState,
-    grid::{to_world_coords, CellType, Grid, GridEntity},
-    map::{ldtk_to_bevy, CollisionMap},
-    screens::state::ScreenState,
-    spritesheet::SpriteSheet,
-    utils::cleanup,
-};
+use crate::prelude::*;
 
 const FLOOR_HEALTH: u32 = 2;
