+++ src/enviro/destructible.rs
@@ -3,14 +3,7 @@ use bevy_bobs::component::health::Health;
 use bevy_ecs_ldtk::EntityInstance;
 use iyes_loopless::prelude::*;
 
-use crate::{
-    constants::BEING_LAYER,
-    grid::{to_world_coords, CellType, GridEntity},
-    map::ldtk_to_bevy,
-    screens::state::ScreenState,
-    spritesheet::{SpriteIndex, SpriteSheet},
-    utils::cleanup,
-};
+use crate::prelude::*;
 
 #[derive(Component)]
 pub struct Destructible;
