+++ src/enviro/dropped_item.rs
@@ -2,12 +2,7 @@ use bevy::prelude::*;
 use bevy_bobs::prefab::{PrefabId, PrefabLib};
 use pino_utils::some_or_return;
 
-use crate::{
-    constants::BEING_LAYER,
-    grid::{to_world_coords, CellType, Grid, GridEntity},
-    item::ItemPrefab,
-    spritesheet::SpriteSheet,
-};
+use crate::{item::ItemPrefab, prelude::*};
 
 #[derive(Component)]
 pub struct DroppedItem {
