+++ src/enviro/water_tile.rs
@@ -3,14 +3,7 @@ use bevy_ecs_ldtk::EntityInstance;
 use iyes_loopless::prelude::*;
 use pino_utils::ok_or_continue;
 
-use crate::{
-    constants::BEING_LAYER,
-    game::GameState,
-    grid::{to_world_coords, CellType, Grid, GridEntity},
-    map::ldtk_to_bevy,
-    spritesheet::{SpriteIndex, SpriteSheet},
-    utils::Dir,
-};
+use crate::prelude::*;
 
 #[derive(Component)]
 pub struct WaterTile {
