+++ src/main.rs
@@ -1,7 +1,6 @@
-#![allow(unused)]
-#![warn(unused_imports)]
+// #![allow(unused)]
+// #![warn(unused_imports)]
 #![allow(dead_code)]
-mod ability;
 mod ai;
 mod app;
 mod attack;
@@ -24,6 +23,24 @@ mod spritesheet;
 mod ui;
 mod utils;
 mod weapon;
+
+pub mod prelude {
+    pub use crate::{
+        attack::{AttackEvent, AttackPattern},
+        constants::*,
+        enemy::{DamageEnemyEvent, DropTable, Enemy, SpawnEnemyEvent},
+        game::{GameState, PauseGame},
+        grid::{snap_to_grid, to_world_coords, CellType, Grid, GridEntity},
+        level::{Level, LevelCleared, LevelFailed},
+        map::{ldtk_to_bevy, CollisionMap, CurrentLevel, SwitchLevelEvent},
+        movement::Movement,
+        player::{DamagePlayerEvent, Player, PlayerMovedEvent, SelectedPlayer, SpawnPlayerEvent},
+        screens::state::ScreenState,
+        spritesheet::*,
+        utils::*,
+        weapon::{CurrentWeapon, Damage, WeaponPrefab},
+    };
+}
 
 fn main() {
     cli::run_cli().unwrap();
