+++ src/player/controller.rs
@@ -6,18 +6,12 @@ use pino_utils::ok_or_return;
 
 use super::{Player, PlayerMovedEvent, SelectedPlayer};
 use crate::{
-    attack::AttackEvent,
     enviro::dropped_item::DroppedItem,
-    game::GameState,
-    grid::{CellType, Grid, GridEntity},
-    movement::Movement,
-    screens::state::ScreenState,
-    spritesheet::SpriteFrames,
+    prelude::*,
     ui::{
         attack_animation::SpawnAttackAnimEvent, attack_indicator::AttackIndicator,
         move_indicator::MoveIndicator,
     },
-    utils::{cardinal_dirs, Dir},
 };
 
 /// Indicator for the current troop action
