+++ src/player/mod.rs
@@ -21,16 +21,8 @@ use self::{
 };
 use crate::{
     camera::CameraFollow,
-    constants::BEING_LAYER,
-    grid::{to_world_coords, CellType, GridEntity},
-    level::Level,
-    map::ldtk_to_bevy,
-    movement::Movement,
-    screens::state::ScreenState,
-    spritesheet::SpriteSheet,
+    prelude::*,
     ui::{attack_indicator::AttackIndicator, move_indicator::MoveIndicator},
-    utils::cleanup,
-    weapon::CurrentWeapon,
 };
 
 /// Tag component for the currently selected troop
