+++ src/player/prefab.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 use bevy_bobs::prefab::{PrefabId, PrefabLib};
 use serde::Deserialize;
 
-use crate::spritesheet::SpriteIndex;
+use crate::prelude::*;
 
 #[derive(Deserialize, PartialEq, Eq, Debug, Clone, Copy)]
 pub enum Class {
